Struct dnssector::compress::Compress

pub struct Compress;

Methods

impl Compress

fn check_compressed_name(packet: &[u8], offset: usize) -> Result<usize>

Checks that an encoded DNS name is valid. This includes following indirections for compressed names, checks for label lengths, checks for truncated names and checks for cycles. Returns the location right after the name.

fn copy_uncompressed_name(
name: &mut Vec<u8>,
packet: &[u8],
offset: usize
) -> UncompressedNameResult

Uncompresses a name starting at offset, and puts the result into name. This function assumes that the input is trusted and doesn't perform any checks. Returns the length of the name as well as the location right after the (possibly compressed) name.

fn uncompress(packet: &[u8]) -> Result<Vec<u8>>