Module dnssector::constants

Enums

Class
DNS query class

EdnsOption
EDNS option

Opcode
DNS opcodes

Rcode
DNS return codes

Section
DNS packet section

Type
DNS query type

Constants

DNS_EDNS_RR_CODE_OFFSET
Offset to the extended RR code, from the beginning of an extended RR.

DNS_EDNS_RR_HEADER_SIZE
Size of the header of an extended RR.

DNS_EDNS_RR_RDLEN_OFFSET
Offset to the length of an extended RR, from the beginning of the extended RR.

DNS_FLAGS_OFFSET
Offset to the flags (including rcode and opcode), from the beginning of the DNS packet

DNS_FLAG_AA

DNS_FLAG_AD

DNS_FLAG_CD

DNS_FLAG_DO

DNS_FLAG_QR

DNS_FLAG_RA

DNS_FLAG_RD

DNS_FLAG_TC

DNS_HEADER_SIZE
Size of the main DNS header, that includes flags and the size of each section.

DNS_MAX_HOSTNAME_INDIRECTIONS
Maximum number of indirections in a compressed name.

DNS_MAX_HOSTNAME_LEN
Maximum length of a host name.

DNS_OPT_RR_EDNS_EXT_FLAGS_OFFSET
Offset to the edns extended flags, for OPT RRs.

DNS_OPT_RR_EDNS_VERSION_OFFSET
Offset to the edns version number, for OPT RRs.

DNS_OPT_RR_EXT_RCODE_OFFSET
Offset to the extended rcode, for OPT RRs.

DNS_OPT_RR_HEADER_SIZE
Size of the header of an edns section.

DNS_OPT_RR_MAX_PAYLOAD_OFFSET
Offset to the maximum payload for edns0-enabled UDP packets, for OPT RRs.

DNS_OPT_RR_RDLEN_OFFSET
Offset to the total size of the edns records, within OPT RRS.

DNS_QUESTION_OFFSET
Offset to the first (and usually only) record of the question section, from the start of the packet.

DNS_RCODE_OFFSET
Offset to the return code, from the beginning of the DNS packet

DNS_RR_CLASS_OFFSET
Offset to the class identifier, from the location right after the RR name.

DNS_RR_HEADER_SIZE
Size of the header for a RR in a section that is not the question section.

DNS_RR_QUESTION_HEADER_SIZE
Size of the header for a RR in the question section.

DNS_RR_RDLEN_OFFSET
Offset to the rdata length, from the location right after the RR name.

DNS_RR_TTL_OFFSET
Offset to the TTL, from the location right after the RR name.

DNS_RR_TYPE_OFFSET
Offset to the type identifier, from the location right after the RR name.

DNS_TID_OFFSET
Offset to the transaction ID, from the beginning of a DNS packet