Struct dnssector::dns_sector::DNSSector

pub struct DNSSector {
    pub packet: Vec<u8>,
    pub offset: usize,
    pub edns_start: Option<usize>,
    pub edns_end: Option<usize>,
    pub edns_count: u16,
    pub ext_rcode: Option<u8>,
    pub edns_version: Option<u8>,
    pub ext_flags: Option<u16>,
}

A DNSSector object summarizes the structure of a DNS packet, so that individual sections can be accessed quickly.

The functions implemented here assume an untrusted input packet.

Fields
packet: Vec<u8> offset: usize edns_start: Option<usize> edns_end: Option<usize> edns_count: u16 ext_rcode: Option<u8> edns_version: Option<u8> ext_flags: Option<u16>
Methods

impl DNSSector

fn qdcount(packet: &[u8]) -> u16

Returns the number of records in the question section.

fn set_qdcount(packet: &mut [u8], value: u16)

Changes the number of questions.

fn ancount(packet: &[u8]) -> u16

Returns the numbersof records in the answer section.

fn set_ancount(packet: &mut [u8], value: u16)

Changes the number of records in the answer section.

fn nscount(packet: &[u8]) -> u16

Returns the number of records in the nameservers section.

fn set_nscount(packet: &mut [u8], value: u16)

Changes the number of records in the nameservers section.

fn arcount(packet: &[u8]) -> u16

Returns the number of records in the additional section.

fn set_arcount(packet: &mut [u8], value: u16)

Changes the number of records in the additional section.

fn set_offset(&mut self, offset: usize) -> Result<usize>

Sets the internal offset to the data to be parsed to an arbitrary location

fn increment_offset(&mut self, n: usize) -> Result<usize>

Increments the internal offset

fn rr_rdlen(&self) -> Result<usize>

Returns the length of the data for the record currently being parsed.

fn new(packet: Vec<u8>) -> Result<Self>

Builds a DNSSector structure for a given untrusted DNS packet.

fn parse(self) -> Result<ParsedPacket>

Parses and validates all records from all sections of an untrusted DNS packet. If the validation succeeds, a ParsedPacket structure containing information to quickly access (extended) flags and individual sections is returned.

fn edns_rr_rdlen(&self) -> Result<usize>

Returns the record length of a record within the edns pseudo-section.

fn check_uncompressed_name(packet: &[u8], offset: usize) -> Result<usize>

Checks that an untrusted encoded DNS name is valid and does not contain any indirections. Returns the location right after the name.