Struct dnssector::parsed_packet::ParsedPacket

pub struct ParsedPacket {
    pub dns_sector: DNSSector,
    pub offset_question: Option<usize>,
    pub offset_answers: Option<usize>,
    pub offset_nameservers: Option<usize>,
    pub offset_additional: Option<usize>,
    pub offset_edns: Option<usize>,
    pub edns_count: u16,
    pub ext_rcode: Option<u8>,
    pub edns_version: Option<u8>,
    pub ext_flags: Option<u16>,
}

A ParsedPacket structure contains information about a successfully parsed DNS packet, that allows quick access to (extended) flags and to individual sections.

Fields
dns_sector: DNSSector offset_question: Option<usize> offset_answers: Option<usize> offset_nameservers: Option<usize> offset_additional: Option<usize> offset_edns: Option<usize> edns_count: u16 ext_rcode: Option<u8> edns_version: Option<u8> ext_flags: Option<u16>
Methods

impl ParsedPacket

fn into_dns_sector(self) -> DNSSector

Converts a ParsedPacket back into a DNSSector.

fn into_packet(self) -> Vec<u8>

Converts a ParsedPacket back into a raw packet.

fn into_iter_question(&mut self) -> Option<QuestionIterator>

Iterates over the question section.

fn into_iter_answer(&mut self) -> Option<AnswerIterator>

Iterates over the answer section.

fn into_iter_nameservers(&mut self) -> Option<NameServersIterator>

Iterates over the list of name servers.

fn into_iter_additional(&mut self) -> Option<AdditionalIterator>

Iterates over the additional section.

fn into_iter_edns(&mut self) -> Option<EdnsIterator>

Iterates over the records from the optional edns pseudo-section.

fn tid(&self) -> u16

Returns the transaction ID.

fn flags(&self) -> u32

Returns the flags, including extended flags. The extended flags optionally obtained using edns are exposed as the highest 16 bits, instead of having distinct sets of flags. The opcode and rcode are intentionally masked in order to prevent misuse: these bits are never supposed to be accessed individually.

fn rcode(&self) -> u8

Returns the return code.

fn opcode(&self) -> u8

Returns the opcode.