Module dnssector::rr_iterator

Structs

RRIterator
An RRIterator structure is a generic way to iterate over the records of a pre-parsed DNS packet. The packet is assumed to have been previously verified for conformance, so the functions provided here are optimized for speed instead of paranoia, and don't return catchable errors: out-of-bounds accesses will make the thread panic, which is exactly what we want: if this ever happens, it means that we failed at properly verifying the packet, so this is a bug, and it has to be fixed, not ignored.

RRRaw
Accessor to the raw packet data. offset is the offset to the current RR. name_end is the offset to the data right after the name.

RRRawMut
Mutable accessor to the raw packet data. offset is the offset to the current RR. name_end is the offset to the data right after the name.

Traits

DNSIterable
The DNSIterable trait represents a set of records that can be iterated over.

RdataIterable

TypedIterable