Struct dnssector::rr_iterator::RRIterator

pub struct RRIterator<'t> {
    pub parsed_packet: &'t mut ParsedPacket,
    pub section: Section,
    pub offset: Option<usize>,
    pub offset_next: usize,
    pub name_end: usize,
    pub rrs_left: u16,
}

An RRIterator structure is a generic way to iterate over the records of a pre-parsed DNS packet. The packet is assumed to have been previously verified for conformance, so the functions provided here are optimized for speed instead of paranoia, and don't return catchable errors: out-of-bounds accesses will make the thread panic, which is exactly what we want: if this ever happens, it means that we failed at properly verifying the packet, so this is a bug, and it has to be fixed, not ignored.

Fields
parsed_packet: &'t mut ParsedPacket section: Section offset: Option<usize> offset_next: usize name_end: usize rrs_left: u16
Methods

impl<'t> RRIterator<'t>

fn new(parsed_packet: &'t mut ParsedPacket, section: Section) -> Self

Creates a new iterator over a pre-parsed packet, for the given section.

fn skip_name(packet: &[u8], offset: usize) -> usize

Quickly skips over a DNS name, without validation/decompression. Returns the location right after the name.

fn skip_rdata(packet: &[u8], offset: usize) -> usize

fn skip_rr(packet: &[u8], offset: usize) -> usize

fn edns_skip_rr(packet: &[u8], offset: usize) -> usize