Trait dnssector::rr_iterator::DNSIterable

pub trait DNSIterable {
    fn next(self) -> Option<Self>
    where
        Self: Sized;
    fn offset(&self) -> Option<usize>;
    fn raw(&self) -> RRRaw;
    fn raw_mut(&mut self) -> RRRawMut;

    fn name_slice(&self) -> &[u8] { ... }
    fn rdata_slice(&self) -> &[u8] { ... }
    fn name_slice_mut(&mut self) -> &mut [u8] { ... }
    fn rdata_slice_mut(&mut self) -> &mut [u8] { ... }
}

The DNSIterable trait represents a set of records that can be iterated over.

Required Methods

fn next(self) -> Option<Self> where
Self: Sized,

Returns the next record, or None if there aren't any left.

fn offset(&self) -> Option<usize>

Returns the offset of the current RR, or None if we haven't started iterating yet.

fn raw(&self) -> RRRaw

Access the raw packet data.

fn raw_mut(&mut self) -> RRRawMut

Access the mutable raw packet data.

Provided Methods

fn name_slice(&self) -> &[u8]

Access the raw packet data, starting from the name.

fn rdata_slice(&self) -> &[u8]

Access the raw packet data, starting from right after the name.

fn name_slice_mut(&mut self) -> &mut [u8]

Access the mutable raw packet data, starting from the name.

fn rdata_slice_mut(&mut self) -> &mut [u8]

Access the mutable raw packet data, starting from right after the name.