Trait dnssector::rr_iterator::DNSIterable

pub trait DNSIterable {
    fn next(self) -> Option<Self>
    where
        Self: Sized;

    fn offset(&self) -> Option<usize>;

    fn offset_next(&self) -> usize;

    fn set_offset(&mut self, offset: usize);

    fn set_offset_next(&mut self, offset: usize);

    fn invalidate(&mut self);

    fn recompute_rr(&mut self);

    fn recompute_sections(&mut self);

    fn raw(&self) -> RRRaw;

    fn raw_mut(&mut self) -> RRRawMut;

    fn parsed_packet(&self) -> &ParsedPacket;

    fn parsed_packet_mut(&mut self) -> &mut ParsedPacket;

    fn is_tombstone(&self) -> bool { ... }

    fn packet(&self) -> &[u8] { ... }

    fn name_slice(&self) -> &[u8] { ... }

    fn rdata_slice(&self) -> &[u8] { ... }

    fn name_slice_mut(&mut self) -> &mut [u8] { ... }

    fn rdata_slice_mut(&mut self) -> &mut [u8] { ... }

    fn uncompress(&mut self) -> Result<()> { ... }
}

The DNSIterable trait represents a set of records that can be iterated over.

Required Methods

fn next(self) -> Option<Self> where
Self: Sized,

Returns the next record, or None if there aren't any left.

fn offset(&self) -> Option<usize>

Returns the offset of the current RR, or None if we haven't started iterating yet or if the current record has been deleted.

In order to check for the later, please use is_tombstone() instead for clarity.

fn offset_next(&self) -> usize

Returns the offset right after the current RR.

fn set_offset(&mut self, offset: usize)

Sets the offset of the current RR.

fn set_offset_next(&mut self, offset: usize)

Sets the offset of the next RR.

fn invalidate(&mut self)

Prevents access to the current record. This is useful after a delete operation: from a user perspective, the current iterator doesn't point to a valid RR any more.

fn recompute_rr(&mut self)

Updates the precomputed RR information

fn recompute_sections(&mut self)

Updates the precomputed offsets of each section.

fn raw(&self) -> RRRaw

Accesses the raw packet data.

fn raw_mut(&mut self) -> RRRawMut

Accesses the mutable raw packet data.

fn parsed_packet(&self) -> &ParsedPacket

Accesses the parsed packet structure.

fn parsed_packet_mut(&mut self) -> &mut ParsedPacket

Accesses the parsed packet structure.

Provided Methods

fn is_tombstone(&self) -> bool

Returns true if the record has been invalidated by a previous call to delete()

fn packet(&self) -> &[u8]

Raw packet data.

fn name_slice(&self) -> &[u8]

Accesses the raw packet data, starting from the name.

fn rdata_slice(&self) -> &[u8]

Access the raw packet data, starting from right after the name.

fn name_slice_mut(&mut self) -> &mut [u8]

Accesses the mutable raw packet data, starting from the name.

fn rdata_slice_mut(&mut self) -> &mut [u8]

Accesses the mutable raw packet data, starting from right after the name.

fn uncompress(&mut self) -> Result<()>

Decompresses the whole packet while keeping the iterator available.